Refactor transfer module and improve test structure
Streamlines the transfer module by removing unused recipient creation functionality and modernizes test implementation:

- Removes deprecated createRecipient method and related types
- Restructures integration tests to use environment variables for credentials
- Improves test assertions and error handling patterns
- Adds more comprehensive test cases for transfer operations
- Cleans up mock implementations by moving them to dedicated files

This change improves code maintainability and testing practices while removing unused functionality.
