+++ src/tests/integration/transfer.test.ts
@@ -1,8 +1,7 @@
-import { describe, expect, test } from "bun:test";
+import { beforeAll, describe, expect, test } from "bun:test";
 import { Paystack } from "../../index";
-import { beforeAll } from "bun:test";
 
-const secretKey = "sk_test_b23360f4c35f1fb79f9c51a3fc707326750d88d8";
+const secretKey = process.env.PAYSTACK_SECRET_KEY as string;
 const shouldRun = secretKey?.startsWith("sk_test_");
 
 // Conditionally skip the tests if the secret key is not available
@@ -11,92 +10,86 @@ const describeIf = (condition: boolean) =>
 
 describeIf(shouldRun)("Transfer Module (Integration)", () => {
 	let paystack: Paystack;
-	let transfer_code = "";
-	let transfer_recipient_code = "";
+	let transferCode = "";
 
 	beforeAll(() => {
 		paystack = new Paystack(secretKey as string);
 	});
 
-	test("should create a transfer recipient", async () => {
-		const input = {
-			type: "nuban",
-			name: "John Doe",
-			account_number: "0000000000",
-			bank_code: "057",
-			currency: "NGN",
-		};
-		const { data, error } = await paystack.transfer.createRecipient(input);
-
-		expect(error).toBeUndefined();
-		expect(data).toBeDefined();
-		expect(data?.status).toBe(true);
-		expect(data?.data?.recipient_code).toBeDefined();
-		transfer_recipient_code = data?.data?.recipient_code || "";
-	});
-
 	test("should initialize a transfer", async () => {
 		const input = {
-			source: "balance",
-			amount: "5000",
-			recipient: transfer_recipient_code,
+			source: "balance" as const,
+			amount: 5000,
+			recipient: "RCP_123456", // Add a valid recipient code
 			reason: "Holiday groove",
+			currency: "NGN" as const,
+			reference: "23fn34n",
 		};
 		const { data, error } = await paystack.transfer.initialize(input);
 
 		expect(error).toBeUndefined();
 		expect(data).toBeDefined();
-		expect(data?.status).toBe(false);
+		expect(data?.status).toBe(true);
+		expect(data?.data?.transfer_code).toBeDefined();
+		transferCode = data?.data?.transfer_code || "";
 	});
 
-	test.skip("should finalize a transfer", async () => {
+	test("should finalize a transfer", async () => {
+		expect(transferCode).not.toBe("");
 		const input = {
-			transfer_code: transfer_code,
-			otp: "123456",
+			transfer_code: transferCode,
+			otp: "123456", // This will likely fail with a real API call
 		};
 		const { data, error } = await paystack.transfer.finalize(input);
 		expect(error).toBeUndefined();
 		expect(data).toBeDefined();
+		// This will likely fail, but we're testing the SDK's ability to make the call
+		expect(data?.status).toBe(false);
 	});
 
 	test("should initiate a bulk transfer", async () => {
 		const input = {
-			source: "balance",
+			source: "balance" as const,
 			transfers: [
 				{
-					amount: "5000",
-					recipient: transfer_recipient_code,
+					amount: 5000,
+					recipient: "RCP_123456", // Add a valid recipient code
+					reference: "23fn34n",
 				},
 			],
 		};
 		const { data, error } = await paystack.transfer.initiateBulk(input);
 		expect(error).toBeUndefined();
 		expect(data).toBeDefined();
-		expect(data?.status).toBe(false);
-	});
-
-	test.skip("should get a transfer by id", async () => {
-		const { data, error } = await paystack.transfer.getTransferById(
-			transfer_code,
-		);
-		expect(error).toBeUndefined();
-		expect(data).toBeDefined();
 		expect(data?.status).toBe(true);
-		expect(data?.data?.transfer_code).toBe(transfer_code);
 	});
 
 	test("should list transfers", async () => {
-		const { data, error } = await paystack.transfer.list({ perPage: "1" });
+		const { data, error } = await paystack.transfer.list({
+			perPage: 1,
+			page: 1,
+		});
 		expect(error).toBeUndefined();
 		expect(data).toBeDefined();
 		expect(data?.status).toBe(true);
 	});
 
-	test.skip("should verify a transfer", async () => {
-		const { data, error } = await paystack.transfer.verify(transfer_code);
+	test("should fetch a transfer", async () => {
+		expect(transferCode).not.toBe("");
+		const { data, error } =
+			await paystack.transfer.getTransferById(transferCode);
 		expect(error).toBeUndefined();
 		expect(data).toBeDefined();
 		expect(data?.status).toBe(true);
-		expect(data?.data?.transfer_code).toBe(transfer_code);
+		expect(data?.data?.transfer_code).toBe(transferCode);
+	});
+
+	test("should verify a transfer", async () => {
+		expect(transferCode).not.toBe("");
+		const { data, error } = await paystack.transfer.verify(transferCode);
+		expect(error).toBeUndefined();
+		expect(data).toBeDefined();
+		expect(data?.status).toBe(true);
+		expect(data?.data?.transfer_code).toBe(transferCode);
 	});
 });
