+++ src/tests/transfer.test.ts
@@ -1,263 +1,174 @@
-import { describe, expect, test } from "bun:test";
-import { Paystack } from "../index";
+import { afterEach, describe, expect, spyOn, test } from "bun:test";
+import { Paystack } from "../../";
 import {
-	transferBulkInitiateSuccess,
-	transferFinalizeSuccess,
-	transferInitiateSuccess,
-	transferListSuccess,
-	transferSingleSuccess,
-} from "../zod/transfer";
-import { mockFetcher } from "./mocks";
+	mockFinalizeTransferResponse,
+	mockInitiateBulkTransferResponse,
+	mockInitiateTransferResponse,
+	mockListTransfersResponse,
+} from "../tests/mocks/transfer";
+import { mockErrorResponse, mockFetch } from "./mocks";
 
-const secretKey = "sk_test_1234567890";
-const paystack = new Paystack(secretKey);
+describe("Transfer Module", () => {
+	const paystack = new Paystack("sk_test_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
 
-describe("Transfer Module (Unit)", () => {
-	test("should initialize a transfer", async () => {
-		const input = {
-			source: "balance",
-			amount: "5000",
-			recipient: "RCP_123456",
-			reason: "Holiday groove",
-		};
-		const expectedData = transferInitiateSuccess.parse({
-			status: true,
-			message: "Transfer has been queued",
-			data: {
-				integration: 100073,
-				domain: "test",
-				amount: 5000,
-				currency: "NGN",
-				source: "balance",
-				reason: "Holiday groove",
-				recipient: 22,
-				status: "pending",
-				transfer_code: "TRF_123456",
-				id: 1,
-				created_at: new Date().toISOString(),
-				updated_at: new Date().toISOString(),
-				reference: "ref_123",
-				titan_code: "titan_123",
-				request: 123,
-				transfersessionid: [],
-				transfertrials: [],
-				failures: null,
-			},
-		});
-
-		paystack.transfer.fetcher = mockFetcher(expectedData, true);
-		const { data, error } = await paystack.transfer.initialize(input);
-		expect(error).toBeUndefined();
-		expect(data).toEqual(expectedData);
+	afterEach(() => {
+		spyOn(global, "fetch").mockRestore();
 	});
 
-	test("should finalize a transfer", async () => {
-		const input = {
-			transfer_code: "TRF_123456",
-			otp: "123456",
-		};
-		const expectedData = transferFinalizeSuccess.parse({
-			status: true,
-			message: "Transfer has been completed",
-			data: {
-				integration: 100073,
-				domain: "test",
-				amount: 5000,
-				currency: "NGN",
-				source: "balance",
-				reason: "Holiday groove",
-				recipient: 22,
-				status: "success",
-				transfer_code: "TRF_123456",
-				id: 1,
-				created_at: new Date().toISOString(),
-				updated_at: new Date().toISOString(),
-				reference: "ref_123",
-				titan_code: "titan_123",
-				request: 123,
-				failures: null,
-				source_details: null,
-			},
-		});
+	describe("initialize", () => {
+		test("should initiate a transfer", async () => {
+			// Arrange
+			const fetchSpy = mockFetch(mockInitiateTransferResponse, true);
+			const input = {
+				source: "balance" as const,
+				amount: 30000,
+				recipient: "RCP_123456",
+				reason: "Calm down",
+				currency: "NGN" as const,
+				reference: "23fn34n",
+			};
 
-		paystack.transfer.fetcher = mockFetcher(expectedData, true);
-		const { data, error } = await paystack.transfer.finalize(input);
-		expect(error).toBeUndefined();
-		expect(data).toEqual(expectedData);
-	});
+			// Act
+			const { data, error } = await paystack.transfer.initialize(input);
 
-	test("should initiate a bulk transfer", async () => {
-		const input = {
-			source: "balance",
-			transfers: [
-				{
-					amount: "5000",
-					recipient: "RCP_123456",
-				},
-			],
-		};
-		const expectedData = transferBulkInitiateSuccess.parse({
-			status: true,
-			message: "2 transfers have been queued",
-			data: [
-				{
-					amount: 20000,
-					recipient: "RCP_123456",
-					reference: "ref_123",
-					currency: "NGN",
-					status: "pending",
-				},
-				{
-					amount: 30000,
-					recipient: "RCP_123456",
-					reference: "ref_456",
-					currency: "NGN",
-					status: "pending",
-				},
-			],
-		});
+			// Assert
+			expect(error).toBeUndefined();
+			expect(data).toBeDefined();
+			expect(data?.status).toBe(true);
+			expect(data?.data?.transfer_code).toBeDefined();
 
-		paystack.transfer.fetcher = mockFetcher(expectedData, true);
-		const { data, error } = await paystack.transfer.initiateBulk(input);
-		expect(error).toBeUndefined();
-		expect(data).toEqual(expectedData);
-	});
-
-	test("should list transfers", async () => {
-		const expectedData = transferListSuccess.parse({
-			status: true,
-			message: "Transfers retrieved",
-			data: [],
-			meta: {
-				total: 1,
-				skipped: 0,
-				perPage: 50,
-				page: 1,
-				pageCount: 1,
-			},
-		});
-
-		paystack.transfer.fetcher = mockFetcher(expectedData, true);
-		const { data, error } = await paystack.transfer.list({});
-		expect(error).toBeUndefined();
-		expect(data).toEqual(expectedData);
-	});
-
-	test("should get a transfer by id", async () => {
-		const expectedData = transferSingleSuccess.parse({
-			status: true,
-			message: "Transfer retrieved",
-			data: {
-				integration: 100073,
-				domain: "test",
-				amount: 5000,
-				currency: "NGN",
-				source: "balance",
-				reason: "Holiday groove",
-				recipient: {
-					domain: "test",
-					type: "nuban",
-					currency: "NGN",
-					name: "John Doe",
-					details: {
-						account_number: "0000000000",
-						account_name: "John Doe",
-						bank_code: "058",
-						bank_name: "Guaranty Trust Bank",
+			// Verify fetch was called correctly
+			expect(fetchSpy).toHaveBeenCalledWith(
+				"https://api.paystack.co/transfer",
+				expect.objectContaining({
+					method: "POST",
+					headers: {
+						"Content-Type": "application/json",
+						Authorization: "Bearer sk_test_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx",
 					},
-					description: "Test recipient",
-					metadata: null,
-					recipient_code: "RCP_123456",
-					active: true,
-					id: 1,
-					integration: 100073,
-					created_at: new Date().toISOString(),
-					updated_at: new Date().toISOString(),
-				},
-				status: "success",
-				transfer_code: "TRF_123456",
-				id: 1,
-				createdAt: new Date().toISOString(),
-				updatedAt: new Date().toISOString(),
-				reference: "ref_123",
-				titan_code: "titan_123",
-				request: 123,
-				failures: null,
-				session: {
-					provider: null,
-					id: null,
-				},
-				fees_charged: 0,
-				fees_breakdown: null,
-				gateway_response: null,
-				source_details: null,
-				created_at: new Date().toISOString(),
-				updated_at: new Date().toISOString(),
-			},
+					body: JSON.stringify(input),
+				}),
+			);
 		});
 
-		paystack.transfer.fetcher = mockFetcher(expectedData, true);
-		const { data, error } = await paystack.transfer.getTransferById("TRF_123456");
-		expect(error).toBeUndefined();
-		expect(data).toEqual(expectedData);
+		test("should handle API error response", async () => {
+			// Arrange
+			mockFetch(mockErrorResponse, false);
+			const input = {
+				source: "balance" as const,
+				amount: 30000,
+				recipient: "RCP_123456",
+				reason: "Calm down",
+				currency: "NGN" as const,
+				reference: "23fn34n",
+			};
+
+			// Act
+			const { data, error } = await paystack.transfer.initialize(input);
+
+			// Assert
+			expect(error).toBeUndefined();
+			expect(data).toBeDefined();
+			expect(data?.status).toBe(false);
+			expect(data?.message).toBe("Invalid key");
+		});
 	});
 
-	test("should verify a transfer", async () => {
-		const expectedData = transferSingleSuccess.parse({
-			status: true,
-			message: "Transfer retrieved",
-			data: {
-				integration: 100073,
-				domain: "test",
-				amount: 5000,
-				currency: "NGN",
-				source: "balance",
-				reason: "Holiday groove",
-				recipient: {
-					domain: "test",
-					type: "nuban",
-					currency: "NGN",
-					name: "John Doe",
-					details: {
-						account_number: "0000000000",
-						account_name: "John Doe",
-						bank_code: "058",
-						bank_name: "Guaranty Trust Bank",
-					},
-					description: "Test recipient",
-					metadata: null,
-					recipient_code: "RCP_123456",
-					active: true,
-					id: 1,
-					integration: 100073,
-					created_at: new Date().toISOString(),
-					updated_at: new Date().toISOString(),
-				},
-				status: "success",
-				transfer_code: "TRF_123456",
-				id: 1,
-				createdAt: new Date().toISOString(),
-				updatedAt: new Date().toISOString(),
-				reference: "ref_123",
-				titan_code: "titan_123",
-				request: 123,
-				failures: null,
-				session: {
-					provider: null,
-					id: null,
-				},
-				fees_charged: 0,
-				fees_breakdown: null,
-				gateway_response: null,
-				source_details: null,
-				created_at: new Date().toISOString(),
-				updated_at: new Date().toISOString(),
-			},
-		});
+	describe("finalize", () => {
+		test("should finalize a transfer", async () => {
+			// Arrange
+			const fetchSpy = mockFetch(mockFinalizeTransferResponse, true);
+			const input = {
+				transfer_code: "TRF_1ptvuv321ahaa7q",
+				otp: "123456",
+			};
 
-		paystack.transfer.fetcher = mockFetcher(expectedData, true);
-		const { data, error } = await paystack.transfer.verify("TRF_123456");
-		expect(error).toBeUndefined();
-		expect(data).toEqual(expectedData);
+			// Act
+			const { data, error } = await paystack.transfer.finalize(input);
+
+			// Assert
+			expect(error).toBeUndefined();
+			expect(data).toBeDefined();
+			expect(data?.status).toBe(true);
+			expect(data?.data?.status).toBe("success");
+
+			// Verify fetch was called correctly
+			expect(fetchSpy).toHaveBeenCalledWith(
+				"https://api.paystack.co/transfer/finalize_transfer",
+				expect.objectContaining({
+					method: "POST",
+					body: JSON.stringify(input),
+				}),
+			);
+		});
+	});
+
+	describe("initiateBulk", () => {
+		test("should initiate a bulk transfer", async () => {
+			// Arrange
+			const fetchSpy = mockFetch(mockInitiateBulkTransferResponse, true);
+			const input = {
+				source: "balance" as const,
+				transfers: [
+					{
+						amount: 20000,
+						recipient: "RCP_123456",
+						reference: "ref_123",
+					},
+					{
+						amount: 30000,
+						recipient: "RCP_123456",
+						reference: "ref_456",
+					},
+				],
+			};
+
+			// Act
+			const { data, error } = await paystack.transfer.initiateBulk(input);
+
+			// Assert
+			expect(error).toBeUndefined();
+			expect(data).toBeDefined();
+			expect(data?.status).toBe(true);
+			expect(data?.data).toBeInstanceOf(Array);
+			expect(data?.data?.length).toBe(2);
+
+			// Verify fetch was called correctly
+			expect(fetchSpy).toHaveBeenCalledWith(
+				"https://api.paystack.co/transfer/bulk",
+				expect.objectContaining({
+					method: "POST",
+					body: JSON.stringify(input),
+				}),
+			);
+		});
+	});
+
+	describe("list", () => {
+		test("should list transfers", async () => {
+			// Arrange
+			const fetchSpy = mockFetch(mockListTransfersResponse, true);
+			const params = { perPage: 50, page: 1 };
+
+			// Act
+			const { data, error } = await paystack.transfer.list(params);
+
+			// Assert
+			expect(error).toBeUndefined();
+			expect(data).toBeDefined();
+			expect(data?.status).toBe(true);
+			expect(data?.data).toBeInstanceOf(Array);
+			expect(data?.meta?.perPage).toBe(50);
+
+			// Verify fetch was called correctly
+			const expectedParams = new URLSearchParams({
+				perPage: "50",
+				page: "1",
+			}).toString();
+			expect(fetchSpy).toHaveBeenCalledWith(
+				`https://api.paystack.co/transfer?${expectedParams}`,
+				expect.any(Object),
+			);
+		});
 	});
 });
